Oops, wrong fix
In the original code there was a comparison if (size >= 0), however GCC
warned that size is a unsigned variable so this comparison is always
true. For some reason, I changed that to if (!size), however this does
not make any sense. Remove the original check since it is always true
anyway.