Modify load_arm9_payload to include offset and max payload size
This is actually a modification made by @d0k3 for his BrahmaLoader
project (that is an actual fork of CakeBrah). It adds two new parameters
for load_arm9_payload function: offset and max payload size.

Offset allows to set an offset for binary payloads, since the majority
of files used as a payload include way more code than the payload itself
(see for example RxTools' code.bin and CakesFW's Cakes.dat). The max
payload size sets the maximum payload size used by the function. This is
maximum size instead of actual size, since the caller generally does not
know the actual size of the payload. This is similar to what main() from
source/main.c does.