Change main() to use load_arm9_payload()
This clean-up the code, avoid an unnecessary memory allocation/copy, and
serves as a example on how to use CakeBrah in other projects.

Tested with Cakes.dat in O3DS with sysNAND 9.2.0-20U. A similar code is
already used and tested in m45t3r/BootCtr@405eb7d24d39757018320c8154ad.